Fix Require cycle: App.js -> components\AppBar.jsx -> screens\HomeScreen.jsx -> App.js
